refactor: update import paths and file names for event repository
- Update import path from "github.com/blackhorseya/monorepo-go/internal/app/domain/event/biz/repo/memory" to "github.com/blackhorseya/monorepo-go/internal/app/domain/event/repo/memory" in file "internal/app/domain/event/biz/impl_external_test.go"
- Rename file "internal/app/domain/event/biz/repo/memory/BUILD.bazel" to "internal/app/domain/event/repo/memory/BUILD.bazel"
- Update import path from "github.com/blackhorseya/monorepo-go/internal/app/domain/event/biz/repo/memory" to "github.com/blackhorseya/monorepo-go/internal/app/domain/event/repo/memory" in file "internal/app/domain/event/biz/repo/memory/memory.go"
- Rename file "internal/app/domain/event/biz/repo/memory/memory_test.go" to "internal/app/domain/event/repo/memory/memory_test.go"
- Update import path from "github.com/blackhorseya/monorepo-go/internal/app/domain/event/biz/repo/memory" to "github.com/blackhorseya/monorepo-go/internal/app/domain/event/repo/memory" in file "internal/app/domain/event/repo/memory/memory_test.go"
