+++ internal/app/domain/event/repo/memory/memory.go
